requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
python-dotenv==1.0.1
numpy==1.26.4 